+++ app/domain/fuel-type.filter.ts
@@ -2,6 +2,7 @@ import * as z from 'zod';
 import { DefaultTake, MaxTake, SortOrder } from './utils';
 
 export enum FuelTypeSortColumns {
+  ORDER = 'order',
   CODE = 'code',
   PRICE_PER = 'pricePer',
   CO2_CONTRIBUTION = 'co2Contribution',
@@ -18,7 +19,7 @@ export const fuelTypeFilterSchema = z
       .default(null),
     skip: z.coerce.number().int().min(0).default(0),
     take: z.coerce.number().int().min(0).max(MaxTake).default(DefaultTake),
-    sortBy: z.enum(Object.values(FuelTypeSortColumns) as [string, ...string[]]).default(FuelTypeSortColumns.CODE),
+    sortBy: z.enum(Object.values(FuelTypeSortColumns) as [string, ...string[]]).default(FuelTypeSortColumns.ORDER),
     sortOrder: z.enum(Object.values(SortOrder) as [string, ...string[]]).default(SortOrder.ASC),
   })
   .strict();